requirements.txt
pprint
gensim
numpy
nltk
langdetect
mecab-python3
scrapy
pymongo
mongoengine
sklearn
matplotlib
simplejson

